refactor: don't expose deprecate as an internal module (#35311)
